Clean up poller test
