+++ node/pkg/watchers/evm/finalizers/neon_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/certusone/wormhole/node/pkg/watchers/evm/connectors"
-	"github.com/certusone/wormhole/node/pkg/watchers/interfaces"
 
 	ethCommon "github.com/ethereum/go-ethereum/common"
 
@@ -19,7 +18,7 @@ import (
 func TestNeonErrorReturnedIfBlockIsNil(t *testing.T) {
 	ctx := context.Background()
 	logger := zap.NewNop()
-	l1Finalizer := interfaces.MockL1Finalizer{LatestFinalizedBlockNumber: 125}
+	l1Finalizer := MockL1Finalizer{LatestFinalizedBlockNumber: 125}
 
 	finalizer := NewNeonFinalizer(logger, &l1Finalizer)
 	assert.NotNil(t, finalizer)
@@ -31,7 +30,7 @@ func TestNeonErrorReturnedIfBlockIsNil(t *testing.T) {
 func TestNeonNotFinalizedIfNoFinalizedL1BlockYet(t *testing.T) {
 	ctx := context.Background()
 	logger := zap.NewNop()
-	l1Finalizer := interfaces.MockL1Finalizer{}
+	l1Finalizer := MockL1Finalizer{}
 
 	finalizer := NewNeonFinalizer(logger, &l1Finalizer)
 	assert.NotNil(t, finalizer)
@@ -50,7 +49,7 @@ func TestNeonNotFinalizedIfNoFinalizedL1BlockYet(t *testing.T) {
 func TestNeonNotFinalizedWhenL1IsLessThanL2(t *testing.T) {
 	ctx := context.Background()
 	logger := zap.NewNop()
-	l1Finalizer := interfaces.MockL1Finalizer{LatestFinalizedBlockNumber: 125}
+	l1Finalizer := MockL1Finalizer{LatestFinalizedBlockNumber: 125}
 
 	finalizer := NewNeonFinalizer(logger, &l1Finalizer)
 	assert.NotNil(t, finalizer)
@@ -69,7 +68,7 @@ func TestNeonNotFinalizedWhenL1IsLessThanL2(t *testing.T) {
 func TestNeonIsFinalizedWhenL1EqualsL2(t *testing.T) {
 	ctx := context.Background()
 	logger := zap.NewNop()
-	l1Finalizer := interfaces.MockL1Finalizer{LatestFinalizedBlockNumber: 125}
+	l1Finalizer := MockL1Finalizer{LatestFinalizedBlockNumber: 125}
 
 	finalizer := NewNeonFinalizer(logger, &l1Finalizer)
 	assert.NotNil(t, finalizer)
@@ -88,7 +87,7 @@ func TestNeonIsFinalizedWhenL1EqualsL2(t *testing.T) {
 func TestNeonIsFinalizedWhenL1GreaterThanL2(t *testing.T) {
 	ctx := context.Background()
 	logger := zap.NewNop()
-	l1Finalizer := interfaces.MockL1Finalizer{LatestFinalizedBlockNumber: 127}
+	l1Finalizer := MockL1Finalizer{LatestFinalizedBlockNumber: 127}
 
 	finalizer := NewNeonFinalizer(logger, &l1Finalizer)
 	assert.NotNil(t, finalizer)
